# src/docs/swagger.yaml
definitions:
...
  github_com_wahid-wex_e-commerce-backend_api_dto.GetOtpRequest:
    properties:
      mobileNumber:
        maxLength: 11
        minLength: 11
        type: string
    required:
    - mobileNumber
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.LoginByUsernameRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_dto.RegisterLoginByMobileRequest:
    properties:
      mobileNumber:
        maxLength: 11
        minLength: 11
        type: string
      otp:
        maxLength: 6
        minLength: 6
        type: string
    required:
    - mobileNumber
    - otp
    type: object
  github_com_wahid-wex_e-commerce-backend_api_dto.RegisterSellerByUsernameRequest:
    properties:
      address:
        minLength: 5
        type: string
      description:
        type: string
      email:
        minLength: 6
        type: string
      logo:
        type: string
      nationalId:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        minLength: 5
        type: string
      storeName:
        minLength: 3
        type: string
      username:
        minLength: 5
        type: string
    required:
    - address
    - nationalId
    - password
    - phone
    - storeName
    - username
    type: object
...
  github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_validation.ValidationError'
        type: array
    type: object
  github_com_wahid-wex_e-commerce-backend_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 42901
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - LimiterError
    - CustomRecovery
    - InternalError
  github_com_wahid-wex_e-commerce-backend_api_validation.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/users/login-seller-by-mobile:
    post:
      consumes:
      - application/json
      description: RegisterLoginSellerByMobileNumber
      parameters:
      - description: RegisterLoginByMobileRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.RegisterLoginByMobileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: RegisterLoginSellerByMobileNumber
      tags:
      - Users
  /v1/users/login-seller-by-username:
    post:
      consumes:
      - application/json
      description: LoginSellerByUsername
      parameters:
      - description: LoginByUsernameRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.LoginByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: LoginSellerByUsername
      tags:
      - Users
...
  /v1/users/refresh-seller-token:
    post:
      consumes:
      - application/json
      description: Send RefreshToken and Then get new token credential
      parameters:
      - description: RefreshTokenRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: Refresh
      tags:
      - Users
...
  /v1/users/register-seller-by-username:
    post:
      consumes:
      - application/json
      description: RegisterSellerByUsername
      parameters:
      - description: RegisterSellerByUsernameRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.RegisterSellerByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: RegisterSellerByUsername
      tags:
      - Users
...
  /v1/users/send-seller-otp:
    post:
      consumes:
      - application/json
      description: Send otp to user
      parameters:
      - description: GetOtpRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_dto.GetOtpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_wahid-wex_e-commerce-backend_api_helper.BaseHttpResponse'
      summary: Send otp to user
      tags:
      - Users
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"